backend/requirements.txt
asgiref==3.6.0
attrs==23.1.0
black==23.3.0
click==8.1.3
colorama==0.4.6
Django==4.2
django-cors-headers==3.14.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
flake8==6.0.0
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.17.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
psycopg2==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
tzdata==2022.6
uritemplate==4.1.1
